Refactoring and extraction of the issueTable
