+++ app/components/Pagination.tsx
@@ -2,7 +2,6 @@
 import { ChevronLeftIcon, ChevronRightIcon, DoubleArrowLeftIcon, DoubleArrowRightIcon } from '@radix-ui/react-icons';
 import { Button, Flex, Text } from '@radix-ui/themes';
 import { useRouter, useSearchParams } from 'next/navigation';
-import React from 'react'
 
 interface Props{
     itemCount: number;
@@ -17,7 +16,7 @@ const Pagination = ({ currentPage, itemCount, pageSize }: Props) => {
 
     const pageCount = Math.ceil(itemCount / pageSize)
     if (pageCount <= 1) return null
-    if (currentPage > pageCount) currentPage = 1
+
 
     const changePage = (page:number) => {
         const params = new URLSearchParams(searchParams)
